refactor : separate file for each function
